include js for iframe auto height
It will be much cleaner if put TABLEs in a separated html document,
and embeds it in markdown, like following:

<iframe src="/wp-content/uploads/yyyy/mm/the-table.xhtml"
frameborder="0" width="100%" scrolling="no"> </iframe>

But the height of the iframe needs to be calculated once
the page is loaded, aka extra JavaScript libraries are imported,
which are "jquery.browser.js" and "jquery-iframe-auto-height.js".

See https://github.com/house9/jquery-iframe-auto-height for detail.