toolchain: CCORI has higher priority than GCC
GCC is only available for CCORI=internal currently.
CCVER is for CCORI != internal.

In the future, GCC and CCVER should be the same logic, but not the same meaning.

CCVER for CCORI not internal.
GCC for CCORI = internal, GCC is the real suffix of /usr/bin/CCPRE-gcc-N, N is the GCC value.

but CCVER may be time or version, blabla.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>