bsp: fix up skip support
The bsp repo is not checked out by default, so, no local master branch
found, use origin/master as the match 'tag'.

It should be there if already downloaded, users should update it
manually instead of always check it, the later is time cost and waste.

To download it manually, just type: 'make bsp'.

Signed-off-by: Wu Zhangjin <falcon@tinylab.org>