nolibc: simplify nolibc program building
Directly include the nolibc.h file, which defines NOLIBC macro for us
and without the requirement of installing the headers.

This built size is a little bigger: 1464, the one built with old method
is only 1440.

To get less size, install the headers and include the required headers
on demand is better.

Signed-off-by: Falcon <falcon@tinylab.org>