cfg: add generic config support
Allow add default config files for all versions, this new config files
are suffixed with .config, which use allnoconfig target instead of
defconfig, it is similar to the kernel xxx.config targets.

This allows us use a minimal configuration without the garbage 'is not
set' lines, please don't run them as 'defconfig' targets.

Signed-off-by: Zhangjin Wu <falcon@tinylab.org>